+++ src/main/java/org/oscwii/shop/controllers/BaseController.java
@@ -1,5 +1,6 @@
 package org.oscwii.shop.controllers;
 
+import org.oscwii.api.OSCAPI;
 import org.oscwii.shop.config.ShopServerConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -7,7 +8,9 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 public abstract class BaseController
 {
     @Autowired
-    private ShopServerConfig config;
+    protected ShopServerConfig config;
+    @Autowired
+    protected OSCAPI api;
 
     @ModelAttribute("handleEc")
     protected boolean handleEc()
