--- src/main/java/org/oscwii/shop/controllers/LandingController.java
@@ -1,20 +1,0 @@
-package org.oscwii.shop.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LandingController extends BaseController
-{
-    @GetMapping("/landing")
-    public String landing()
-    {
-        return "landing";
-    }
-
-    @GetMapping("/home")
-    public String home()
-    {
-        return "home";
-    }
-}
